Frontend: useRequest hook

+++ server/Data/BookingRepository.cs
@@ -37,10 +37,15 @@ namespace Data
                                             .Where(x => x.UserEmail.Equals(userEmail)).ToListAsync();
         }
 
-        public async Task DeleteBooking(string userEmail, int timeSlotId)
+        public async Task<int> DeleteBooking(string userEmail, int timeSlotId)
         {
             Booking booking = await _context.Bookings.FirstOrDefaultAsync(b => b.UserEmail.Equals(userEmail)
                                                                           && b.TimeSlotId == timeSlotId);
+            if (booking == null)
+            {
+                return 
+            }
+
             _context.Bookings.Remove(booking);
 
             await _context.SaveChangesAsync();
